requirements/data.txt
pandas==2.2.2
numpy==1.26.4
tqdm==4.66.4
pillow==10.3.0
opencv-python-headless==4.9.0.80
pyarrow==16.1.0
plotly==5.22.0
streamlit==1.36.0